Port Shadowlings Contracts to EIP-7702
This PR ports the contracts that originally worked with (the now
defunct) EIP-3074 to use EIP-7702. This was a fairly sizable change
as the `Shadowlings` contract went from acting like an "entrypoint"
that controlled the individual accounts to being a delegation target
holding the implementation for the individual accounts. In particular,
that means that we no longer need to restrict `missingPrefund == 0`
because the ERC-4337 `EntryPoint` calls the shadow account directly.

I also started cleaning up the documentation. Note that **the
dockerized setup and app still do not work!** The required changes for
the dockerized setup and app will come in a later PR.

The ZK circuits did not require any changes.

**Notes to testers**: do NOT use Foundry v1.2, it does not currently
validate EIP-7702 nonces correctly.