Use Full Commit for Signature S-Value
In PREP (to name the "Nick's Method" trick we are using here), you
encode some context in the signature: one of the components is some
fixed value (here our `_R`), and the other is a hash of the full
context. For signatures to be valid, you need two things:
- `r` to be a valid x-coordinate of a point on the `secp256k1` curve. It
  turns out that this is ~1/2 of possible 32-bit values, and this is why
  we fix `r` instead of `s`: so we can choose one that we _know_ is
  valid (and why we use `keccak256(...) - 1`, because just
  `keccak256(...)` was not a valid x-coordinate).
- And [specifically for
  EIP-7702](https://eips.ethereum.org/EIPS/eip-7702#behavior), `s` needs
  to be a scalar in the bottom half of the field. So, given
  `N_secp256k1` as the order of the `secp256k1` curve (i.e. the number
  of distinct points you get from summing the generator point with
  itself), then `s < N_secp256k1/2`.

Now, `commit` is the MiMC hash on the `BN254` curve (also referred to as
`bn128` or `alt_bn128` - so many names, don't ask me why). This means
that `commit` will be a scalar of the prime field modulo the order of
the `BN254` curve (i.e. given `N_bn254` the order of the `BN254` curve,
we have `commit < N_bn254`). It just so happens that `N_bn254 =
21888242871839275222246405745257275088696311157297823662689037894645226208583
< N_secp256k1/2 =
57896044618658097711785492504343953926418782139537452191302581570759080747168`.
This means that _all commits_ are valid `s` components for `secp256k1`
signatures, and we don't need to modify `commit` at all for our PREP
scheme 🎉!

So, back to the original question - why `yParity == 0`? Well, we need a
value for `yParity` and we can either:
- Set it to some arbitrary fixed value (either 0 or 1, it doesn't matter
  as long as it is always the same) like we do for `r`
- Have it be generated from the PREP context like we do for `s`

Since we already used all available entropy from `commit` in `s`, it
doesn't really make sense to have `yParity` come from the context hash
(here `commit`), and instead we just choose to fix it to some arbitrary
value - here 0 (since `PUSH0` is slightly better than `PUSH1 0x01`).