Improve 4337 Compatibility (#6)
The generated verifyer contracts are not compatible with 4337 because of
their use of `GAS` opcode when calling pairing precompiles. It looks
like it was setup this way for gas estimation reasons, but it does not
seem to affect our uses (all the tools and scripts run as expected).

With this change, I can execute a 4337 user operation without running
the bundler in `--unsafe` mode.

The specific bundler rule that is being violated is
[`OP-012`](https://ercs.ethereum.org/ERCS/erc-7562#opcode-rules):

> `GAS` (`0x5A`) opcode is allowed, but only if followed immediately by
`*CALL` instructions, else it is blocked. This is a common way to pass
all remaining gas to an external call, and it means that the actual
value is consumed from the stack immediately and cannot be accessed by
any other opcode.